Refactor project structure: reorganize and split folders for better organization
