+++ trip-expense-front/src/components/DestinationCard/DestinationCard.jsx
