+++ trip-expense-front/src/components/HeroSection/HeroSection.css
@@ -3,7 +3,7 @@
     max-width: 1380px;
     width: 100%;
     height: 600px;  
-    background: url("../assets/HeroImg.png") center/cover no-repeat;
+    background: url("../../assets/HeroImg.png") center/cover no-repeat;
     display: flex;
     justify-content: center;
     align-items: center;
