+++ trip-expense-front/src/components/HeroSection/HeroSection.jsx
