+++ trip-expense-front/src/pages/loginPage/LoginPage.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
-import "./LoginPage.css";
-
+import "../loginPage/LoginPage.css"
 const LoginPage = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
