+++ trip-expense-front/src/pages/signUpPage/SignUpPage.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import "./SignUpPage.css";
-import image1 from "../assets/image1.jpg";
-import image2 from "../assets/image2.jpg";
-import image3 from "../assets/image3.jpg";
+import image1 from "../../assets/image1.jpg";
+import image2 from "../../assets/image2.jpg";
+import image3 from "../../assets/image3.jpg";
 
 const images = [image1, image2, image3];
